pyproject.toml
[project]
name = "hbapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pillow>=12.1.0",
    "pytest>=9.0.2",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
